requirements.txt
selenium==4.21.0
undetected-chromedriver==3.5.5
python-dotenv==1.0.1
pandas==2.2.2
openai==1.30.1
httpx==0.27.0
urllib3==2.2.1
pyperclip
webdriver-manager
pyautogui
pillow
